Update Windows Update.ps1

+++ Script Files/Windows Update.ps1
@@ -18,6 +18,26 @@ if ($w64 -or $w32)
 # Check if CMD is Running, Stop Windows Terminal if Running
     if((get-process "cmd" -ea SilentlyContinue) -eq $Null){ 
         echo "" 
+#requires -version 5.1
+#Calling Powershell as Admin and setting Execution Policy to Bypass to avoid Cannot run Scripts error
+if (-NOT ([Security.Principal.WindowsPrincipal][Security.Principal.WindowsIdentity]::GetCurrent()).IsInRole([Security.Principal.WindowsBuiltInRole] "Administrator"))  
+{  
+#Is Powershell 7 Installed
+  $w64=Get-ItemProperty "HKLM:\Software\Wow6432Node\Microsoft\Windows\CurrentVersion\Uninstall\*" | Where-Object { try { $_.DisplayName -match "PowerShell 7-x64" } catch { $false } }
+  $w32=Get-ItemProperty "HKLM:\Software\Microsoft\Windows\CurrentVersion\Uninstall\*"  | Where-Object { try { $_.DisplayName -match "PowerShell 7-x64" } catch { $false } }
+if ($w64 -or $w32)
+{
+  Start-Process pwsh.exe -Verb RunAs -ArgumentList ('-noprofile -noexit -file "{0}" -elevated' -f ($myinvocation.MyCommand.Definition))
+# Check if Windows Terminal is Running, Stop Windows Terminal if Running
+    if((get-process "WindowsTerminal" -ea SilentlyContinue) -eq $Null){ 
+        echo "" 
+    }
+    else{ 
+    Stop-Process -processname "WindowsTerminal"
+        }
+# Check if CMD is Running, Stop Windows Terminal if Running
+    if((get-process "cmd" -ea SilentlyContinue) -eq $Null){ 
+        echo "" 
     }
     else{ 
     Stop-Process -processname "cmd"
@@ -56,6 +76,9 @@ Else{
   Break
     }
 }
+
+Clear-Host
+#Requires -RunAsAdministrator
 
 # Run PSWindowsUpdate
